-5/9/2020 -- Version 0.8.8.4 ---             -Steam approved for release!             -All stats are working (Times escaped, best time, worst time, least amount of clicks to escape)             -All achievement images are in Steam, for the 17 achievements             -Added the secret achievement channel             -Added the WASD achievement             -Added the Mo clicks Mo problems achievement             -Added the interesting puzzle for the egg

+++ steamworks/interfaces/userstats.py
@@ -110,7 +110,7 @@ class SteamUserStats(object):
         :return:
         """
         if isinstance(value, float):
-            return self.steam.SetStatFloat(name, value)
+            return self.steam.SetStatFloat(name, c_float(value))
 
         elif isinstance(value, int):
             return self.steam.SetStatInt(name, value)
